Atividade 12

+++ src/main/java/br/com/ifba/usuario/repository/UsuarioRepository.java
@@ -2,7 +2,7 @@ package br.com.ifba.usuario.repository;
 
 import br.com.ifba.usuario.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+//Interface UsuarioRepository. Tem conexão com o banco de dados (Nesse caso, PostgreSQL).
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
 }
